+++ yandex/data_source_yandex_mdb_mongodb_cluster.go
@@ -229,6 +229,27 @@ func dataSourceYandexMDBMongodbCluster() *schema.Resource {
 				Set:      schema.HashString,
 				Computed: true,
 			},
+			"maintenance_window": {
+				Type:     schema.TypeList,
+				MaxItems: 1,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"type": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"day": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"hour": {
+							Type:     schema.TypeInt,
+							Computed: true,
+						},
+					},
+				},
+			},
 		},
 	}
 }
@@ -350,6 +371,11 @@ func dataSourceYandexMDBMongodbClusterRead(d *schema.ResourceData, meta interfac
 	}
 
 	if err := d.Set("security_group_ids", cluster.SecurityGroupIds); err != nil {
+		return err
+	}
+
+	mw := flattenMongoDBMaintenanceWindow(cluster.MaintenanceWindow)
+	if err := d.Set("maintenance_window", mw); err != nil {
 		return err
 	}