+++ yandex/data_source_yandex_mdb_mongodb_cluster_test.go
@@ -65,6 +65,9 @@ func testAccDataSourceMDBMongoDBClusterAttributesCheck(datasourceName string, re
 			"sharded",
 			"cluster_config.0.version",
 			"security_group_ids",
+			"maintenance_window.0.type",
+			"maintenance_window.0.day",
+			"maintenance_window.0.hour",
 		}
 
 		for _, attrToCheck := range instanceAttrsToTest {
@@ -97,6 +100,9 @@ func testAccDataSourceMDBMongoDBClusterCheck(datasourceName string, resourceName
 		resource.TestCheckResourceAttr(datasourceName, "host.#", "2"),
 		testAccCheckCreatedAtAttr(datasourceName),
 		resource.TestCheckResourceAttr(datasourceName, "security_group_ids.#", "1"),
+		resource.TestCheckResourceAttr(datasourceName, "maintenance_window.0.type", "WEEKLY"),
+		resource.TestCheckResourceAttr(datasourceName, "maintenance_window.0.day", "FRI"),
+		resource.TestCheckResourceAttr(datasourceName, "maintenance_window.0.hour", "20"),
 	)
 }